refactor(ui-overlays): extract create rename bulk-rename trash and restore renderers
